ai_modules/department_mapping-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4
langchain-openai
